pass options --nprint_filter & --nprint_filter_help along to nPrint v1.1.4
resolves #27
