+++ src/nprintml/__init__.py
@@ -1,3 +1,3 @@
 __version__ = '0.0.3'
 
-__nprint_version__ = '1.1.3'
+__nprint_version__ = '1.1.4'
